Move Agent stats to its own file
No changes, just bringing down the size of the Agent
